requirements.txt
pip==21.3.1
click==8.0.4
python-dotenv==0.20.0